pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "laut"
dynamic = ["version"]
description = "Nix build trace signature tool"
requires-python = ">=3.8"
# TODO: add license and classifiers
dependencies = [
    "rfc8785",
    "pyjwt",
    "cryptography",
    "boto3",
    "click",
]

[project.scripts]
laut = "laut.cli:main"
"laut.py" = "laut.cli:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["laut"]

[tool.setuptools_scm]

#[tool.pytest.ini_options]
#testpaths = ["tests"]
#python_files = ["test_*.py", "*_test.py"]
#addopts = "-ra -q"

[tool.coverage.run]
source = ["laut"]
branch = true

[tool.pytest.ini_options]
pythonpath = [
  "."
]
addopts = [
    "--import-mode=importlib",
]

[tool.pyright]
typeCheckingMode = "basic"
